fix: move sanitize_for_export to common_tools to avoid cyclic import
csv_builder and export_builder import each other, so placing the
function in common_tools (which both already depend on) breaks the
cycle.
