+++ onadata/libs/utils/common_tools.py
@@ -26,6 +26,20 @@ from onadata.libs.utils.common_tags import ATTACHMENTS
 DEFAULT_UPDATE_BATCH = 100
 TRUE_VALUES = ["TRUE", "T", "1", 1]
 
+_FORMULA_PREFIXES = ("=", "+", "-", "@", "\t", "\r")
+
+
+def sanitize_for_export(value):
+    """Prevent CSV/XLSX formula injection (CWE-1236).
+
+    Prefixes cell values starting with formula characters with a single quote
+    to force spreadsheet applications to treat them as literal strings.
+    Reference: https://owasp.org/www-community/attacks/CSV_Injection
+    """
+    if isinstance(value, str) and value and value[0] in _FORMULA_PREFIXES:
+        return "'" + value
+    return value
+
 
 class FilenameMissing(Exception):
     """Custom Exception for a missing filename."""
